refactor: rearranging assets for templates
