+++ next-docs/public/templates/handshake.png
