+++ next-docs/public/templates/promo.png
