+++ next-docs/public/templates/sportsbet/partnership.png
